Use WorkManager in place of Firebase Job Dispatcher
Use WorkManager's support for content URI listening to replace use of Firebase Job Dispatcher.
