+++ build.gradle
@@ -27,11 +27,11 @@ buildscript {
         lifecycleVersion = "1.1.1"
         roomVersion = "1.1.1-rc1"
         pagingVersion = "1.0.0"
+        workManagerVersion = "1.0.0-alpha02"
         playServicesWearableVersion = "15.0.1"
         firebaseCoreVersion = "16.0.0"
         firebasePerfVersion = "16.0.0"
         crashlyticsVersion = "2.9.2"
-        jobDispatcherVersion = "0.8.5"
         wearableVersion = "2.3.0"
         okhttpVersion = "3.10.0"
         picassoVersion = "2.71828"
